Create main.c
In this program there are two tasks, one is called simple_task that just send something into the serial line and second task that make the soc go to deep sleep mode.
The idle task send also the tick number. Every time when the button pc13 is pushed, the soc go to deep sleep, otherwise it wakes up.